+++ plugins/iamauth.js
@@ -1,4 +1,4 @@
-// Copyright © 2017, 2019 IBM Corp. All rights reserved.
+// Copyright © 2017, 2021 IBM Corp. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -66,9 +66,7 @@ class IAMPlugin extends BasePlugin {
     delete req.url;
     req.uri = u.format(new u.URL(req.uri), {auth: false});
 
-    self._tokenManager.renewIfRequired().then(() => {
-      callback(state);
-    }).catch((error) => {
+    self._tokenManager.renewIfRequired().catch((error) => {
       debug(error);
       if (state.attempt < state.maxAttempt) {
         state.retry = true;
@@ -85,8 +83,7 @@ class IAMPlugin extends BasePlugin {
       } else {
         state.abortWithResponse = [ error ]; // return error to client
       }
-      callback(state);
-    });
+    }).finally(() => callback(state));
   }
 
   onResponse(state, response, callback) {